Introducing Babel & Webpack
To use EcmaScript 6 syntax, but produce EcmaScript 5 syntax in the end, we use Babel to "transpile" our source code to the ES5 compliant code (https://babeljs.io/).

For the build process we use https://webpack.js.org/. The config can be found in the `webpack.config.js` file.

We basically have an entry point of all our JS files, at least for now and the different loaders keep care of all the dependencies we have inside the files (e.g. for CSS we require in there or other JS files we use the babel-loader to use Babel.js in the end).

With that we have a build process in place to write small JS scripts and compile them together as needed.We furthermore produce sourcemaps which bring the benefit that we can still debug the source code in the browser even if the files landing in the www/ folder are compressed/minified and hardly readable.